bug fixes
—traffic light addition state reset
—traffic light addition mechanism fixed now so all lanes can have
lights added

+++ src/model/TrafficLight.java
@@ -75,7 +75,7 @@ public class TrafficLight {
 		this.status = status;
 	}
 
-	public void setRedInterval(long interval) {
+	public void setRedInterval(float interval) {
 		this.redInterval = interval;
 	}
